# paas-core/apps/api/docs/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "MyPaaS API",
        "description": "Multi-tenant PaaS platform API with billing (Xendit)",
        "version": "0.1.0",
        "contact": {
            "name": "MyPaaS Team"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local development"
        }
    ],
    "paths": {
        "/healthz": {
            "get": {
                "summary": "Health check",
                "tags": [
                    "Health"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/readyz": {
            "get": {
                "summary": "Readiness check",
                "tags": [
                    "Health"
                ],
                "responses": {
                    "200": {
                        "description": "Ready"
                    },
                    "503": {
                        "description": "Not ready"
                    }
                }
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "summary": "Register user",
                "tags": [
                    "Auth"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "minLength": 8
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created"
                    },
                    "409": {
                        "description": "Email exists"
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "summary": "Login",
                "tags": [
                    "Auth"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "JWT token returned"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/v1/auth/refresh": {
            "post": {
                "summary": "Refresh token",
                "tags": [
                    "Auth"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "New token pair"
                    }
                }
            }
        },
        "/api/v1/user/me": {
            "get": {
                "summary": "Get current user",
                "tags": [
                    "User"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User profile"
                    }
                }
            },
            "put": {
                "summary": "Update current user",
                "tags": [
                    "User"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile updated"
                    }
                }
            }
        },
        "/api/v1/orgs": {
            "get": {
                "summary": "List organizations",
                "tags": [
                    "Organizations"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of orgs"
                    }
                }
            },
            "post": {
                "summary": "Create organization",
                "tags": [
                    "Organizations"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "name",
                                    "slug"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "slug": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Organization created"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}